refactor: Destination.kt 제거 및 BookingCompleteActivity 뒤로가기 버튼 로직 수정
- MovieBookingActivity 에 대한 프레젠터, 뷰 제거
- Destination 제거
