+++ app/src/androidTest/java/woowacourse/movie/ui/main/MovieBookingActivityTest.kt
@@ -9,6 +9,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
+import woowacourse.movie.ui.MovieBookingActivity
 
 class MovieBookingActivityTest {
     @get:Rule
