+++ app/src/androidTest/java/woowacourse/movie/ui/movielist/MovieListFragmentTest.kt
@@ -8,7 +8,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
-import woowacourse.movie.ui.main.MovieBookingActivity
+import woowacourse.movie.ui.MovieBookingActivity
 
 class MovieListFragmentTest {
     @get:Rule
