+++ app/src/main/java/woowacourse/movie/ui/booking/view/BookingActivity.kt
@@ -9,6 +9,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.databinding.DataBindingUtil
+import java.time.LocalDate
+import java.time.LocalTime
 import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivityBookingBinding
 import woowacourse.movie.domain.model.Headcount
@@ -24,8 +26,6 @@ import woowacourse.movie.ui.booking.view.spinner.listener.ScreeningTimeListener
 import woowacourse.movie.ui.seat.BookingSeatActivity
 import woowacourse.movie.utils.StringFormatter
 import woowacourse.movie.utils.intentSerializable
-import java.time.LocalDate
-import java.time.LocalTime
 
 class BookingActivity :
     AppCompatActivity(),
@@ -112,6 +112,7 @@ class BookingActivity :
                 theaterName,
             ),
         )
+        finish()
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean =
