--- app/src/main/java/woowacourse/movie/utils/Destination.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.utils
-
-enum class Destination {
-    HISTORY,
-    COMPLETE,
-    SEAT,
-    SETTINGS,
-    HOME,
-}
