+++ app/src/main/res/layout/activity_movie_booking.xml
@@ -7,7 +7,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.MovieBookingActivity">
+        tools:context=".ui.MovieBookingActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/main_fragment_container_view"
